refactor: move form class to the end in blade templates
This commit moves the action form class to the end of the blade templates. The class was previously inline with the button, which made it difficult to style. By moving the class to the end, it is now easier to style and does not affect the functionality of the form.