feat: implement server side datatable
This commit implements a server-side data table for the attendance report page.

The old data table was not able to handle large amounts of data efficiently, and it was also not able to be sorted or filtered by the user. The new server-side data table addresses these issues by loading data from the server only when it is needed, and by providing the ability to sort and filter the data.

BREAKING CHANGE: The old data table has been removed and replaced with the new server-side data table. This may cause some minor changes to the appearance of the attendance report page, but it should not affect the functionality of the page.