refactor: rename bank-account to customer
Rename all bank-account related files to customer

This commit renames all bank-account related files to customer. This includes renaming the model, controller, views, routes, and migrations.

This change was made to better reflect the purpose of these files, which is to manage customer data. The term "bank-account" is no longer used in the application, so it is no longer necessary to use it in the file names.