feat: implement billing search feature
This commit implements a billing search feature that allows users to search for billings by destination and customer name.

The search feature is implemented by adding a search parameter to the BillingController index method. The search parameter is then used to filter the billing results.

The search feature is a valuable addition to the billing system as it allows users to quickly and easily find the billing information they need.