feat: add bank management feature
This commit adds a new feature that allows users to manage banks. This includes creating, editing, and deleting banks, as well as viewing a list of all banks.

This feature will be useful for businesses that need to track their bank accounts and transactions. It will also be useful for individuals who want to keep track of their personal finances.