refactor: update user relationship and fetch customer billing follow ups in officer report
This commit refactors the user relationship and adds a new query to fetch customer billing follow ups in the officer report.

The previous implementation of the user relationship was not optimal and caused performance issues. This commit updates the relationship to use eager loading, which will improve performance.

Additionally, this commit adds a new query to fetch customer billing follow ups in the officer report. This query will allow officers to see a list of all customer billing follow ups that have been created in the specified date range.

BREAKING CHANGE: The old user relationship is no longer supported. Any code that relies on the old relationship will need to be updated to use the new relationship.