fix(excel-by-user): convert pdf to image
This commit converts the pdf to image for the signature_officer and signature_customer columns in the officer-reports/excel-by-user view.

The previous implementation was using the <img> tag to display the pdf, which was not working correctly. This change converts the pdf to an image using the base64 encoding, which is then displayed in the <img> tag.

This change also includes a fix for an issue where the images were not being displayed correctly when using Safari.